+++ derive_builder/src/lib.rs
@@ -216,6 +216,31 @@
 //! }
 //! ```
 //!
+//! ## Fallible Setters
+//!
+//! Alongside the normal setter methods, you can expose fallible setters which are generic over
+//! the `TryInto` trait. TryInto is a not-yet-stable trait 
+//! (see rust-lang issue [#33417](https://github.com/rust-lang/rust/issues/33417)) similar to
+//! `Into` with the key distinction that the conversion can fail, and therefore produces a `Result`.
+//! 
+//! You can declare the `try_setter` attribute today, and your builder will work on stable Rust -
+//! the fallible setters will automatically light up in future when the `try_from` feature is
+//! stabilized.
+//!
+//!
+//! ```rust,ignore
+//! # #[macro_use]
+//! # extern crate derive_builder;
+//! #
+//! #[derive(Builder, Debug, PartialEq)]
+//! #[builder(try_setter, setter(into))]
+//! struct Lorem {
+//!     pub ipsum: u8,
+//! }
+//! ```
+//!
+//! A complete example of using fallible setters is available in `examples/try_setter.rs`.
+//!
 //! ## Default Values
 //!
 //! You can define default values for each field via annotation by `#[builder(default="...")]`,
@@ -366,6 +391,9 @@
 //! - If derive_builder depends on your crate, and vice versa, then a cyclic
 //!   dependency would occur. To break it you could try to depend on the
 //!   [`derive_builder_core`] crate instead.
+//! - The `try_setter` attribute and `owned` builder pattern are not compatible in practice;
+//!   an error during building will consume the builder, making it impossible to continue
+//!   construction.
 //!
 //! ## Debugging Info
 //!